Make further changes/fixes to reading in file
Program can now read in a midi file to completion.
Needs testing to see if it formats the data correctly, but the program
terminates.

All data from midi events is now read into a char array, as in most
cases, the values we get are in bytes.
Also, we now have some enums associated with all track events.