Data is finally read correctly, and the program prints out midi notes.
Turns out that the one big thing that I was missing was something called
running status, where events without a status byte >= 0x80 should be
assumed to have the same status byte as the previous event.

Also, I was reading data in backwards. I was assuming that all bytes
would be reversed in order due to the endianness, but didn't realize
that only values > 1 byte would be formatted differently.